Catch Lua unittests runtime errors Signed-off-by: Sergey Matveev <stargrave@stargrave.org>

+++ remotetesting/lib/functions.sh
@@ -71,9 +71,12 @@ run_lua()
 	retreive_lua $luasrc_path | prepare_lua > $luasrc
 	CMD_SCP "$luasrc" /tmp/"$luasrc_name"
 	CMD_SCP $WORK_DIR/lib/luaunit.lua /usr/lib/lua/luaunit.lua
-	CMD_SSH PATH=/bin:/sbin:/usr/bin:/usr/sbin lua /tmp/"$luasrc_name"
+	local run_result=`mktemp`
+	CMD_SSH PATH=/bin:/sbin:/usr/bin:/usr/sbin lua /tmp/"$luasrc_name" |
+		tee $run_result
+	grep -q "^Failed" $run_result && return 1 || true
 	CMD_SSH rm -f /tmp/"$luasrc_name"
-	rm -f $luasrc
+	rm -f $luasrc $run_result
 }
 
 run_clearing()
